+++ customer_portal/src/epilot/sdkconfiguration.py
@@ -20,9 +20,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '1.0.0'
-    sdk_version: str = '2.1.3'
+    sdk_version: str = '2.1.4'
     gen_version: str = '2.173.0'
-    user_agent: str = 'speakeasy-sdk/python 2.1.3 2.173.0 1.0.0 epilot-customer-portal'
+    user_agent: str = 'speakeasy-sdk/python 2.1.4 2.173.0 1.0.0 epilot-customer-portal'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> Tuple[str, Dict[str, str]]: